feat(power-assert-formatter): add support for yield statements
From what I can tell, there is no way to get esprima to parse a
yield statement that is not wrapped in a generator function.
To work around this: try to parse normally first, then try
parsing by wrapping with a generator statement. If it works
the second time, then we assume it is due to a yield statement.